Changes in input.cfg:
added:
- output_tag_time    1
- mc_output_folder   output and other mc parameters
- llg_max_walltime        0:0:0 and equivalents
- gneb_output_energies_step             0   - gneb_output_energy was removed
- gneb_output_energies_interpolated     1
- gneb_output_energies_divide_by_nspins 1
- gneb_output_chain_step 0
renamed:
- llg_output_energy_single is now llg_output_energy_step
- llg_output_configuration_single -> ..._step

+++ core/src/Spirit/Configurations.cpp
@@ -332,7 +332,7 @@ void Configuration_Hopfion(State *state, float r, int order, const float positio
 	Vector3 vpos = image->geometry->center + _pos;
 
 	// Set cutoff radius
-	if (r_cut_spherical < 0) r_cut_spherical = r*M_PI;
+	if (r_cut_spherical < 0) r_cut_spherical = (float)r*M_PI;
 
 	// Create position filter
 	auto filter = get_filter(vpos, r_cut_rectangular, r_cut_cylindrical, r_cut_spherical, inverted);
